requirements.txt
dspy-ai>=2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pydantic>=2.0.0
pytest>=7.4.0
rich>=13.0.0
